% Generated by build script
\name{cl_courts}
\alias{cl_courts}
\title{List courts available on CourtListener}
\usage{
cl_courts()
}
\arguments{
\item{page_size}{Number of results (default 20)}
}
\value{tibble: id, full_name, short_name, jurisdiction, url}
\description{List courts available on CourtListener}
