% Generated by build script
\name{cl_opinion}
\alias{cl_opinion}
\title{Fetch a CourtListener opinion cluster by ID}
\usage{
cl_opinion()
}
\arguments{
\item{id}{Opinion cluster ID (integer)}
}
\value{tibble: one row with id, caseName, court, dateFiled, snippet, url, plain_text}
\description{Fetch a CourtListener opinion cluster by ID}
