% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{cl_opinion}
\alias{cl_opinion}
\title{Fetch a CourtListener opinion cluster by ID}
\usage{
cl_opinion(id)
}
\arguments{
\item{id}{Opinion cluster ID (integer)}
}
\value{
tibble: one row with id, caseName, court, dateFiled, snippet, url, plain_text
}
\description{
Fetch a CourtListener opinion cluster by ID
}
